pyproject.toml
[tool.poetry]
name = "presentable"
version = "0.1.0"
description = "Presentable Object Model system"
authors = ["Your Name"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
lark-parser = "^0.12.0"
confuse = "^2.0.1"
pyyaml = "^6.0.2"
setuptools = "^77.0.3"
lark = "^1.2.2"
weasyprint = "^65.1"
html-to-markdown = "^1.3.2"
premailer = "^3.10.0"
diagrams = "^0.24.4"
pymupdf = "^1.25.5"
pymupdf4llm = "^0.0.22"
mdformat = "^0.7.22"
beautifulsoup4 = "^4.13.4"
inflect = "^7.5.0"
markdown = "^3.8"
pandas = "^2.2.3"
plantuml = "^0.3.0"
plantweb = "^1.3.0"
emojis = "^0.7.0"
emoji = "^2.14.1"
pydantic = "^2.11.5"
jsonpickle = "^4.1.1"
deepdiff = "^8.5.0"
emojipedia = "^0.4.1"
openpyxl = "^3.1.5"
jsonschema = "^4.24.0"
cairosvg = "^2.8.2"
docraptor = "^3.1.0"
playwright = "^1.53.0"
pyprince = "^0.6.1"
mermaid-py = "^0.8.0"
pikepdf = "^9.9.0"
mermaid-cli = "^0.1.2"
chromium = "^0.0.0"
networkx = "^3.5"
anthropic = "^0.58.2"
playwright-stealth = "^2.0.0"
2captcha-python = "^1.5.1"
zenrows = "^1.4.0"
openai = "^1.97.0"
google-genai = "^1.26.0"
llm-openrouter = "^0.4.1"
openrouter = "^1.0"
rdflib = "^7.1.4"
n3 = "^0.2.0"
typing-inspection = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
